requirements.txt
Django==1.11
django-environ==0.4.3
django-fsm==2.5.0
django-phonenumber-field==1.3.0
djangorestframework-jwt==1.10.0
djangorestframework==3.6.2
psycopg2==2.7.1
boto3==1.4.4
gunicorn==19.7.1
factory-boy==2.8.1
